pyproject.toml
[project]
name = "confluence-publisher"
version = "0.1.0"
description = "A local command-line tool for publishing Markdown files to Atlassian Confluence with automatic conversion to Confluence-compatible format and support for creating subpages"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mistune>=3.1.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
